+++ KernelBench/level1/50_conv_standard_2D__square_input__square_kernel.py
@@ -5,7 +5,7 @@ import torch.nn.functional as F
 class Model(nn.Module):
     def __init__(self, num_classes=1000):
         super(Model, self).__init__()
-        self.conv1 = nn.Conv2d(in_channels=3, out_channels=96, kernel_size=11, stride=4, padding=2)
+        self.conv1 = nn.Conv2d(in_channels=3, out_channels=96, kernel_size=11, stride=4, padding=2, dtype=torch.bfloat16)
     
     def forward(self, x):
         x = self.conv1(x)
@@ -16,7 +16,7 @@ batch_size = 256
 num_classes = 1000
 
 def get_inputs():
-    return [torch.rand(batch_size, 3, 224, 224)]
+    return [torch.rand(batch_size, 3, 224, 224, dtype=torch.bfloat16)]
 
 def get_init_inputs():
     return [num_classes]
