+++ KernelBench/level1/59_conv_standard_3D__asymmetric_input__square_kernel.py
@@ -17,7 +17,7 @@ class Model(nn.Module):
     """
     def __init__(self, in_channels: int, out_channels: int, kernel_size: int, stride: int = 1, padding: int = 0, dilation: int = 1, groups: int = 1, bias: bool = False):
         super(Model, self).__init__()
-        self.conv3d = nn.Conv3d(in_channels, out_channels, (kernel_size, kernel_size, 1), stride=stride, padding=padding, dilation=dilation, groups=groups, bias=bias)
+        self.conv3d = nn.Conv3d(in_channels, out_channels, (kernel_size, kernel_size, 1), stride=stride, padding=padding, dilation=dilation, groups=groups, bias=bias, dtype=torch.bfloat16)
         
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
@@ -41,7 +41,7 @@ height = 256
 depth = 10
 
 def get_inputs():
-    x = torch.rand(batch_size, in_channels, height, width, depth)
+    x = torch.rand(batch_size, in_channels, height, width, depth, dtype=torch.bfloat16)
     return [x]
 
 def get_init_inputs():
