+++ KernelBench/level1/68_conv_transposed_3D__square_input__asymmetric_kernel.py
@@ -18,7 +18,7 @@ class Model(nn.Module):
     """
     def __init__(self, in_channels: int, out_channels: int, kernel_size: tuple, stride: tuple = (1, 1, 1), padding: tuple = (0, 0, 0), output_padding: tuple = (0, 0, 0), groups: int = 1, bias: bool = False):
         super(Model, self).__init__()
-        self.conv_transpose3d = nn.ConvTranspose3d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, output_padding=output_padding, groups=groups, bias=bias)
+        self.conv_transpose3d = nn.ConvTranspose3d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, output_padding=output_padding, groups=groups, bias=bias, dtype=torch.bfloat16)
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
@@ -44,7 +44,7 @@ width = 64
 height = 64
 
 def get_inputs():
-    x = torch.rand(batch_size, in_channels, depth, width, height)
+    x = torch.rand(batch_size, in_channels, depth, width, height, dtype=torch.bfloat16)
     return [x]
 
 def get_init_inputs():
